% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordiselect.r
\name{ordiselect}
\alias{ordiselect}
\title{Select species for ordination plots}
\usage{
ordiselect(matrix, ord, ablim = 1, fitlim = 1, choices = c(1, 2),
  method = "axes", env, freq = FALSE)
}
\arguments{
\item{matrix}{Community data, a matrix-like object with samples in rows and species in columns.}

\item{ord}{\code{vegan} ordination result object (e.g. from \code{\link[vegan]{decorana}}, \code{\link[vegan]{cca}}, \code{\link[vegan]{rda}} or \code{\link[vegan]{metaMDS}}).}

\item{ablim}{Proportion of species with highest abundances displayed.}

\item{fitlim}{Proportion of species with best fit displayed.}

\item{choices}{Axes shown.}

\item{method}{The species fit method: \code{"axes"} or \code{"vars"}.}

\item{env}{Fitted environmental variabes (minimum = 2 variables, must be result object of \code{\link[vegan]{envfit}}). Only used if \code{method = "vars"}.}

\item{freq}{It this is set on \code{TRUE}, the frequency of species is used instead of cover-abundance values.}
}
\description{
This function simplifies the selection of relevant species in ordination diagrams. It works with results from the \code{vegan} package. The selection can be based upon cover abundance values and/or species fit.
The resulting object contains the names of the selected species and can be used for the \code{select} argument in ordination plots.
}
\section{Details}{

Two methods for species fit are implemented.
\itemize{\item In \code{method = "axes"} species scores are used for fitting. This is the default method.
\item In (\code{method = "vars"}) environmental variables are used for fitting. Here the Euclidean distance between species and environment variable centroids is calculated.}
If \code{method = "vars"} is used, the environmental variables need to be fitted with \code{\link[vegan]{envfit}} and the result of this function must be provided to the \code{env} argument.
The function uses only significant variables (p < 0.05).
If there are less than \emph{two} significant variables, only cover abundance values are used and a warning message is displayed.

The two described methods work well both in eigenvalue-based and in distance-based ordinations.
If axes fit should be applied on distance-based ordination, species scores need to be calculated during the analysis, e.g. by selecting \code{wascores = TRUE} in \code{\link[vegan]{metaMDS}}.
However, contrary to eigenvalue-based methods axes do not have a direct meaning in distance-based methods.
Alternatively a fit to the environmental variables is provided.

If no limit is defined for one of the arguments all species are displayed.
}
\examples{
## Calculate DCA
library(vegan)
scheden.dca <- decorana(schedenveg)

## Select the 30\% most abundant species and call the result
limited <- ordiselect(schedenveg, scheden.dca, ablim = 0.3)
limited

## Use the result in plotting
plot(scheden.dca, display="n")
points(scheden.dca, display="sites")
points(scheden.dca, display="species",
   select = limited, pch=3, col="red", cex=0.7)
ordipointlabel(scheden.dca, display="species",
   select = limited, col="red", cex=0.7, add = TRUE)

## Select the 30\% most frequent species with 50\% best axis fit
limited <- ordiselect(schedenveg, scheden.dca, ablim = 0.3,
   fitlim = 0.5, freq = TRUE)

## Select the 30\% most abundant species with 60\% best environmental fit
## in NDMS for axes 1 & 3
nmds <- metaMDS(schedenveg, k = 3)   # run NMDS
env13 <- envfit(nmds, schedenenv[,2:10], choices=c(1,3))
limited13 <- ordiselect(schedenveg, nmds, ablim = 0.3, fitlim = 0.6,
   choices = c(1,3), method = "vars", env = env13)
}
\author{
Friedemann Goral \email{fgoral@gwdg.de} and Jenny Schellenberg
}

