% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedenveg.r
\docType{data}
\name{schedenveg}
\alias{schedenveg}
\title{Vegetation releves from Scheden}
\format{
A data frame with 28 rows (samples) and 155 variables (species)
}
\usage{
data(schedenveg)
}
\description{
An example vegetation dataset containing 28 grassland releves from Scheden, Niedersachsen, Germany.
The releves were done May 2016 during a students field course at the University of Goettingen.
Locations at the study site are based on the diploma thesis from \cite{Eichholz (1997)}
}
\references{
Eichholz, A. (1997): Wiesen und Magerrasen am Suedhang des Hohen Hagen. Diplomarbeit Biologie, University of Goettingen.
}
\keyword{datasets}
