% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synsort_long.R
\name{synsort_long}
\alias{synsort_long}
\title{Sorting functions for synoptic tables (long format)}
\usage{
synsort_long(
  syn1,
  syn2 = syn1,
  db,
  cluster,
  min1 = 0,
  min2 = 0
)
}
\arguments{
\item{syn1}{Input synoptic table 1, usually from \code{\link{syntable_long}}.}
\item{syn2}{Optional second input table with additional numeric sorting criteria.}
\item{db}{Community data in long format. The first column must contain sample identifiers, the second column species names and the third column numeric values.}
\item{cluster}{Integer or character vector/factor with classification cluster identity matching the samples.}
\item{min1}{Cluster-wise threshold minimum value for species shown in the final sorted synoptic table.}
\item{min2}{Threshold minimum value for considering species values of a numerical second input table \code{syn2}.}
}
\value{Returns an (invisible) list analogous to \code{\link{synsort}}.}
\description{
This function sorts synoptic tables from \code{\link{syntable_long}} output. Sorting criteria can be defined by one or two numeric input tables. Only the \code{"allspec"} method is implemented.
}
\seealso{\code{\link{syntable_long}}, \code{\link{synsort}}}
