% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntable_long.R
\name{syntable_long}
\alias{syntable_long}
\title{Synoptic tables for long-format vegetation databases}
\usage{
syntable_long(db, cluster, abund = "percentage", type = "percfreq", digits = 0)
}
\arguments{
\item{db}{Data frame in long format containing at least the columns
`RELEVE_NR`, `TaxonName` and `Cover_Perc`.}

\item{cluster}{Integer or character vector/factor with classification
cluster identity for each `RELEVE_NR`.}

\item{abund}{Type of abundances. Define whether `Cover_Perc` contains
percentage cover (`abund = "percentage"`, default) or presence/absence
data (`abund = "pa"`, with values 0/1).}

\item{type}{Type of synoptic table output
`type = c("percfreq", "totalfreq", "mean", "median", "phi")`.}

\item{digits}{Integer indicating the number of decimal places to be
displayed in result tables (default 0).}
}
\value{
An (invisible) list with components:
  \item{\code{$syntable }}{unordered synoptic table for given species and clusters}
  \item{\code{$samplesize }}{total number of samples per cluster}
}
\description{
`syntable_long` calculates synoptic tables for long-format vegetation data
where species occur in rows and samples are given by `RELEVE_NR`.
}
\seealso{
[syntable()], [synsort()]
}
