% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_matrix.R
\name{trans_matrix}
\alias{trans_matrix}
\title{Transpose species matrix}
\usage{
trans_matrix(matrix, row.names = T)
}
\arguments{
\item{matrix}{Community data, a data frame.}

\item{row.names}{A logical evaluation indicating whether the new column names are stored as row names of the data frame \code{TRUE} \emph{(default)} or in the first column \code{FALSE}.}
}
\value{
A transposed data frame.
}
\description{
The function transposes a species matrix, while preserving correct species and sample names. 
The new column names must be stored as row names of the data frame. They may also be stored in the first column, when chosing the argument \code{row.names = F}.
}
\examples{
# Transpose species matrix
schedenveg.trans <- trans_matrix(schedenveg)

}
\author{
Friedemann von Lampe (\email{fvonlampe@uni-goettingen.de})
}
