% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimcheckMDS.R
\name{dimcheckMDS}
\alias{dimcheckMDS}
\title{Plot the stress values of NMDS for tested dimensions}
\usage{
dimcheckMDS(matrix, distance = "bray", k = 6, trymax = 20)
}
\arguments{
\item{matrix}{Community data, a matrix-like object with samples in rows and species in columns.}

\item{distance}{Dissimilarity index used in vegdist.}

\item{k}{Number of dimensions.}

\item{trymax}{Maximum number of random starts in search of stable solution.}
}
\description{
This function provides a simple plot of stress values for a given number of tested dimensions (default \code{k = 6}) in NMDS.
It is based on function \code{\link[vegan]{metaMDS}} (\code{vegan} package) and uses the \code{monoMDS} engine.
}
\section{Details}{

The plot shows the border of the 0.2 stress value limit.
\cite{Clarke 1993} suggests the following guidelines for acceptable stress values:
<0.05 = excellent, <0.10 = good, <0.20 = usable, >0.20 = not acceptable.
}
\examples{
## Use of function with default values
dimcheckMDS(meadows)

## Use of function for testing 10 dimensions
dimcheckMDS(meadows, k = 10)
}
\author{
Jenny Schellenberg and Friedemann Goral \email{fgoral@gwdg.de}
}
\references{
Clarke, K. R. (1993). Non-parametric multivariate analysis of changes in community structure. \emph{Austral J Ecol} \strong{18:} 117-143.
}
\seealso{
\code{\link[vegan]{metaMDS}} \code{\link[vegan]{stressplot}}
}

