% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racurve.r
\name{racurve}
\alias{racurve}
\title{Draw rank-abundance curves}
\usage{
racurve(matrix, main = "Rank-abundance diagram", ylog = FALSE)
}
\arguments{
\item{matrix}{Community data, a matrix-like object with samples in rows.}

\item{main}{The main title.}

\item{ylog}{If set on \code{TRUE} the y-axis is displayed on a log-scale.}
}
\description{
This function draws a rank-abundance curve for community data.
If you wish to draw multiple rank-abundance curves for selected samples use \link{racurves}.
}
\examples{
## Draw simple rank-abundance curve
racurve(meadows)

## Draw simple rank-abundance curve with log-scaled axis
racurve(meadows, ylog=T)

## Draw multiple rank-abundance curves for selected samples
racurve(meadows, samples=c(22,31,45))
}
\author{
Friedemann Goral (\email{fgoral@gwdg.de})
}
\seealso{
\code{\link{racurves}}
}

