% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racurves.r
\name{racurves}
\alias{racurves}
\title{Draw multiple rank-abundance curves for selected samples}
\usage{
racurves(matrix, main = "Rank-abundance diagram", bw = TRUE)
}
\arguments{
\item{matrix}{Community data, a matrix-like object with samples in rows and species in columns. Rank-abundance curves are drawn for all selected rows (samples).}

\item{main}{The main title.}

\item{bw}{If set on \code{FALSE} the lines will be drawn in colours instead of black/white lines with different line types.}
}
\description{
This function draws multiple rank-abundance curves for selected samples into one diagram.
If you wish to draw a simple rank-abundance curve see \link{racurve}.
}
\section{Details}{

Rank abundance curves or Whittaker plots (see \cite{Whittaker 1965}) are used to display relative species abundance as biodiversity component.
They are a means to visualise species richness and species evenness.

The axes of the diagram will be scaled according automatically.
As the line type is used to differenciate between samples, a maximum of 6 curves per diagram is feasible.
}
\examples{
## Draw multiple rank-abundance curves for selected samples
racurves(meadows[c(22,31,45), ]
}
\author{
Friedemann Goral \email{fgoral@gwdg.de}
}
\references{
Whittaker, R. H. (1965). Dominance and Diversity in Land Plant Communities: Numerical relations of species express the importance of competition in community function and evolution. \emph{Science} \strong{147 :} 250-260.
}
\seealso{
\code{\link{racurve}}
}

