% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racurves.r
\name{racurves}
\alias{racurves}
\title{Draw multiple rank-abundance curves for selected samples}
\usage{
racurves(matrix, main = "Rank-abundance diagram", bw = TRUE)
}
\arguments{
\item{matrix}{Community data, a matrix-like object with samples in rows and species in columns. Rank-abundance curves are drawn for all selected rows (samples).}

\item{main}{The main title.}

\item{bw}{If set on \code{FALSE} the lines will be drawn in colours instead of black/white lines with different line types.}
}
\description{
This function draws multiple rank-abundance curves for selected samples into one diagram.
If you wish to draw a simple rank-abundance curve see \link{racurve}.
}
\section{Details}{

The axes of the diagram will be scaled according automatically.
As the line type is used to differenciate between samples, a maximum of 6 curves per diagram is feasible.
}
\examples{
## Draw simple rank-abundance curve
racurve(meadows)

## Draw simple rank-abundance curve with log-scaled axis
racurve(meadows, ylog=T)

## Draw multiple rank-abundance curves for selected samples
racurve(meadows[c(22,31,45), ]
}
\author{
Friedemann Goral (\email{fgoral@gwdg.de})
}
\seealso{
\code{\link{racurve}}
}

