% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_schedenenv.r
\docType{data}
\name{schedenenv}
\alias{schedenenv}
\title{Header data for Vegetation releves from Scheden}
\format{
A data frame with 28 rows (samples) and 10 variables
 \itemize{
  \item{comm: }{Plant community as defined in 1997: \emph{Arrhenatheretum} or \emph{Gentiano-Koelerietum}}
  \item{altit: }{Altitude (m)}
  \item{north: }{North value as cosine of aspect}
  \item{slope: }{Slope (degrees)}
  \item{cov_herb: }{Cover of herb layer (\%)}
  \item{cov_litt: }{Cover of litter (\%)}
  \item{cov_moss: }{Cover of mosses (\%)}
  \item{cov_opensoil: }{Cover of open soil (\%)}
  \item{height_herb: }{Average height of herb layer (cm)}
  \item{soil_depth: }{Soil depth (cm)}
}
}
\usage{
schedenenv
}
\description{
An example vegetation dataset containing 28 grassland releves from Scheden, Niedersachsen, Germany.
The releves were done May 2016 during a students field course at the University of Goettingen.
Locations at the study site are based on the diploma thesis from \cite{Eichholz (1997)}
}
\references{
Eichholz, A. (1997): Wiesen und Magerrasen am Suedhang des Hohen Hagen. Diplomarbeit Biologie, University of Goettingen.
}
\keyword{datasets}
